% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOMWidget.R
\name{jupyter.widget.DOMWidget}
\alias{jupyter.widget.DOMWidget}
\alias{DOMWidget}
\title{generate the mime bundle for the widget}
\usage{
DOMWidget(..., error_call = current_env())
}
\arguments{
\item{...}{See constructor for \code{jupyter.widgets.DOMWidget}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
a DOM Widget
}
\section{Super class}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{jupyter.widget.DOMWidget}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{style}}{the \code{\link[=Style]{Style()}} for this widget}

\item{\code{layout}}{the \code{\link[=Layout]{Layout()}} for this widget}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tabbable}}{Is this widget tabbable. Read/Write}

\item{\code{tooltip}}{tooltip. Read/Write}

\item{\code{_dom_classes}}{DOM classes. Read only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.DOMWidget-new}{\code{jupyter.widget.DOMWidget$new()}}
\item \href{#method-jupyter.widget.DOMWidget-mime_bundle}{\code{jupyter.widget.DOMWidget$mime_bundle()}}
\item \href{#method-jupyter.widget.DOMWidget-clone}{\code{jupyter.widget.DOMWidget$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="observe"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-observe'><code>jupyter.widgets.base::jupyter.widget.Widget$observe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on'><code>jupyter.widgets.base::jupyter.widget.Widget$on()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.DOMWidget-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.DOMWidget-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.DOMWidget$new(
  layout = Layout(),
  style = NULL,
  tabbable = FALSE,
  tooltip = "",
  `_dom_classes` = character(),
  `_model_module` = "@jupyter-widgets/base",
  `_model_module_version` = "2.0.0",
  `_model_name` = "",
  `_view_module` = "@jupyter-widgets/base",
  `_view_count` = NULL,
  `_view_module_version` = "2.0.0",
  `_view_name` = "",
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout}}{a \code{\link[=Layout]{Layout()}}}

\item{\code{style}}{a \code{\link[=Style]{Style()}}}

\item{\code{tabbable}}{is the widget tabbable}

\item{\code{tooltip}}{tooltip}

\item{\code{_dom_classes}}{CSS classes applied to widget DOM element}

\item{\code{_model_module}}{The namespace of the model.}

\item{\code{_model_module_version}}{A semver requirement for namespace version containing the model.}

\item{\code{_model_name}}{model name}

\item{\code{_view_module}}{view mmodule}

\item{\code{_view_count}}{view count}

\item{\code{_view_module_version}}{view module version}

\item{\code{_view_name}}{view name}

\item{\code{...}}{unused}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.DOMWidget-mime_bundle"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.DOMWidget-mime_bundle}{}}}
\subsection{Method \code{mime_bundle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.DOMWidget$mime_bundle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the mime bundle for the widget
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.DOMWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.DOMWidget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.DOMWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
