% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{factory}
\alias{factory}
\title{Factory function for R6 jupyter classes}
\usage{
factory(class)
}
\arguments{
\item{class}{R6 class object}
}
\value{
a function that takes \code{...} and \code{error_call} and forwards
them to the constructor of the class.
}
\description{
Factory function for R6 jupyter classes
}
