% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_Style.R
\name{jupyter.widget.Style}
\alias{jupyter.widget.Style}
\title{Sytle class}
\description{
Sytle class

Sytle class
}
\section{Super class}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{jupyter.widget.Style}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.Style-new}{\code{jupyter.widget.Style$new()}}
\item \href{#method-jupyter.widget.Style-clone}{\code{jupyter.widget.Style$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_custom"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_custom'><code>jupyter.widgets.base::jupyter.widget.Widget$on_custom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_update'><code>jupyter.widgets.base::jupyter.widget.Widget$on_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Style-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Style-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Style$new(
  `_model_module` = "@jupyter-widgets/base",
  `_model_module_version` = "2.0.0",
  `_model_name` = "",
  `_view_module` = "@jupyter-widgets/base",
  `_view_count` = NULL,
  `_view_module_version` = "2.0.0",
  `_view_name` = "",
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{_model_module}}{The namespace of the model.}

\item{\code{_model_module_version}}{A semver requirement for namespace version containing the model.}

\item{\code{_model_name}}{model name}

\item{\code{_view_module}}{view mmodule}

\item{\code{_view_count}}{view count}

\item{\code{_view_module_version}}{view module version}

\item{\code{_view_name}}{view name}

\item{\code{...}}{unused}

\item{\code{error_call}}{see \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Style-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Style-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Style$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
