% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Widget.R
\name{set_widget_state_check}
\alias{set_widget_state_check}
\title{Set a state checker}
\usage{
set_widget_state_check(class, name, fun)
}
\arguments{
\item{class}{widget class}

\item{name}{name of the state}

\item{fun}{function that takes a single parameter, checks it and returns something that is suitable for a state}
}
\description{
Set a state checker
}
