% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{update_list}
\alias{update_list}
\alias{default_or}
\alias{null_or}
\alias{arg_match_or_empty}
\alias{ensure}
\title{update list}
\usage{
update_list(x, ...)

default_or(default, fun)

null_or(fun)

arg_match_or_empty(
  arg,
  values = NULL,
  ...,
  multiple = FALSE,
  error_arg = caller_arg(arg),
  error_call = caller_env()
)

ensure(x, fun = assertthat::is.string, ..., msg = NULL)
}
\arguments{
\item{x}{list to update}

\item{...}{new items}

\item{default}{default value}

\item{fun}{function}

\item{arg}{A symbol referring to an argument accepting strings.}

\item{values}{A character vector of possible values that \code{arg} can take.}

\item{multiple}{Whether \code{arg} may contain zero or several values.}

\item{error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{msg}{a custom error message to be printed if one of the conditions is
false.}
}
\description{
update list
}
