% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unbox_one_of}
\alias{unbox_one_of}
\title{Create a state check function for a given set of accepted values}
\usage{
unbox_one_of(values, allow_empty = FALSE, allow_null = FALSE)
}
\arguments{
\item{values}{accepted values}

\item{allow_empty}{TRUE if "" is accepted}

\item{allow_null}{TRUE if NULL is accepted}
}
\description{
Create a state check function for a given set of accepted values
}
