% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Layout.R
\name{Layout}
\alias{Layout}
\title{Layout}
\usage{
Layout(
  align_content = NULL,
  align_items = NULL,
  align_self = NULL,
  border_bottom = NULL,
  border_left = NULL,
  border_right = NULL,
  border_top = NULL,
  bottom = NULL,
  display = NULL,
  flex = NULL,
  flex_flow = NULL,
  grid_area = NULL,
  grid_auto_columns = NULL,
  grid_auto_flow = NULL,
  grid_auto_rows = NULL,
  grid_column = NULL,
  grid_gap = NULL,
  grid_row = NULL,
  grid_template_areas = NULL,
  grid_template_columns = NULL,
  grid_template_rows = NULL,
  height = NULL,
  justify_content = NULL,
  justify_items = NULL,
  left = NULL,
  margin = NULL,
  max_height = NULL,
  max_width = NULL,
  min_height = NULL,
  min_width = NULL,
  object_fit = NULL,
  object_position = NULL,
  order = NULL,
  overflow = NULL,
  padding = NULL,
  right = NULL,
  top = NULL,
  visibility = NULL,
  width = NULL,
  ...,
  error_call = caller_env()
)
}
\arguments{
\item{align_content}{The align-content CSS attribute.}

\item{align_items}{The align-items CSS attribute.}

\item{align_self}{The align-self CSS attribute.}

\item{border_bottom}{The border bottom CSS attribute.}

\item{border_left}{The border left CSS attribute.}

\item{border_right}{The border right CSS attribute.}

\item{border_top}{The border top CSS attribute.}

\item{bottom}{The bottom CSS attribute.}

\item{display}{The display CSS attribute.}

\item{flex}{The flex CSS attribute.}

\item{flex_flow}{The flex-flow CSS attribute.}

\item{grid_area}{The grid-area CSS attribute.}

\item{grid_auto_columns}{The grid-auto-columns CSS attribute.}

\item{grid_auto_flow}{The grid-auto-flow CSS attribute.}

\item{grid_auto_rows}{The grid-auto-rows CSS attribute.}

\item{grid_column}{The grid-column CSS attribute.}

\item{grid_gap}{The grid-gap CSS attribute.}

\item{grid_row}{The grid-row CSS attribute.}

\item{grid_template_areas}{The grid-template-areas CSS attribute.}

\item{grid_template_columns}{The grid-template-columns CSS attribute.}

\item{grid_template_rows}{The grid-template-rows CSS attribute.}

\item{height}{The height CSS attribute.}

\item{justify_content}{The justify-content CSS attribute.}

\item{justify_items}{The justify-items CSS attribute.}

\item{left}{The left CSS attribute.}

\item{margin}{The margin CSS attribute.}

\item{max_height}{The max-height CSS attribute.}

\item{max_width}{The max-width CSS attribute.}

\item{min_height}{The min-height CSS attribute.}

\item{min_width}{The min-width CSS attribute.}

\item{object_fit}{The object-fit CSS attribute.}

\item{object_position}{The object-position CSS attribute.}

\item{order}{The order CSS attribute.}

\item{overflow}{The overflow CSS attribute.}

\item{padding}{The padding CSS attribute.}

\item{right}{The right CSS attribute.}

\item{top}{The top CSS attribute.}

\item{visibility}{The visibility CSS attribute.}

\item{width}{The width CSS attribute.}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a new \code{jupyter.widget.Layout} object
}
\description{
Layout
}
