% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layout.R
\name{jupyter.widget.Layout}
\alias{jupyter.widget.Layout}
\alias{Layout}
\title{Layout class}
\usage{
Layout(..., error_call = current_env())
}
\arguments{
\item{...}{See constructor for the \code{jupyter.widgets.Layout} class}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Layout class

Layout class
}
\section{Super class}{
\code{\link[jupyter.widgets.base:jupyter.widget.Widget]{jupyter.widgets.base::jupyter.widget.Widget}} -> \code{jupyter.widget.Layout}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-jupyter.widget.Layout-new}{\code{jupyter.widget.Layout$new()}}
\item \href{#method-jupyter.widget.Layout-clone}{\code{jupyter.widget.Layout$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="check_state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-check_state'><code>jupyter.widgets.base::jupyter.widget.Widget$check_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_custom"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_custom'><code>jupyter.widgets.base::jupyter.widget.Widget$on_custom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="on_update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-on_update'><code>jupyter.widgets.base::jupyter.widget.Widget$on_update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="set_state_check"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-set_state_check'><code>jupyter.widgets.base::jupyter.widget.Widget$set_state_check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="state"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-state'><code>jupyter.widgets.base::jupyter.widget.Widget$state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jupyter.widgets.base" data-topic="jupyter.widget.Widget" data-id="update"><a href='../../jupyter.widgets.base/html/jupyter.widget.Widget.html#method-jupyter.widget.Widget-update'><code>jupyter.widgets.base::jupyter.widget.Widget$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Layout-new"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Layout-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Layout$new(
  align_content = NULL,
  align_items = NULL,
  align_self = NULL,
  border_bottom = NULL,
  border_left = NULL,
  border_right = NULL,
  border_top = NULL,
  bottom = NULL,
  display = NULL,
  flex = NULL,
  flex_flow = NULL,
  grid_area = NULL,
  grid_auto_columns = NULL,
  grid_auto_flow = NULL,
  grid_auto_rows = NULL,
  grid_column = NULL,
  grid_gap = NULL,
  grid_row = NULL,
  grid_template_areas = NULL,
  grid_template_columns = NULL,
  grid_template_rows = NULL,
  height = NULL,
  justify_content = NULL,
  justify_items = NULL,
  left = NULL,
  margin = NULL,
  max_height = NULL,
  max_width = NULL,
  min_height = NULL,
  min_width = NULL,
  object_fit = NULL,
  object_position = NULL,
  order = NULL,
  overflow = NULL,
  padding = NULL,
  right = NULL,
  top = NULL,
  visibility = NULL,
  width = NULL,
  ...,
  error_call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{align_content}}{The align-content CSS attribute.}

\item{\code{align_items}}{The align-items CSS attribute.}

\item{\code{align_self}}{The align-self CSS attribute.}

\item{\code{border_bottom}}{The border bottom CSS attribute.}

\item{\code{border_left}}{The border left CSS attribute.}

\item{\code{border_right}}{The border right CSS attribute.}

\item{\code{border_top}}{The border top CSS attribute.}

\item{\code{bottom}}{The bottom CSS attribute.}

\item{\code{display}}{The display CSS attribute.}

\item{\code{flex}}{The flex CSS attribute.}

\item{\code{flex_flow}}{The flex-flow CSS attribute.}

\item{\code{grid_area}}{The grid-area CSS attribute.}

\item{\code{grid_auto_columns}}{The grid-auto-columns CSS attribute.}

\item{\code{grid_auto_flow}}{The grid-auto-flow CSS attribute.}

\item{\code{grid_auto_rows}}{The grid-auto-rows CSS attribute.}

\item{\code{grid_column}}{The grid-column CSS attribute.}

\item{\code{grid_gap}}{The grid-gap CSS attribute.}

\item{\code{grid_row}}{The grid-row CSS attribute.}

\item{\code{grid_template_areas}}{The grid-template-areas CSS attribute.}

\item{\code{grid_template_columns}}{The grid-template-columns CSS attribute.}

\item{\code{grid_template_rows}}{The grid-template-rows CSS attribute.}

\item{\code{height}}{The height CSS attribute.}

\item{\code{justify_content}}{The justify-content CSS attribute.}

\item{\code{justify_items}}{The justify-items CSS attribute.}

\item{\code{left}}{The left CSS attribute.}

\item{\code{margin}}{The margin CSS attribute.}

\item{\code{max_height}}{The max-height CSS attribute.}

\item{\code{max_width}}{The max-width CSS attribute.}

\item{\code{min_height}}{The min-height CSS attribute.}

\item{\code{min_width}}{The min-width CSS attribute.}

\item{\code{object_fit}}{The object-fit CSS attribute.}

\item{\code{object_position}}{The object-position CSS attribute.}

\item{\code{order}}{The order CSS attribute.}

\item{\code{overflow}}{The overflow CSS attribute.}

\item{\code{padding}}{The padding CSS attribute.}

\item{\code{right}}{The right CSS attribute.}

\item{\code{top}}{The top CSS attribute.}

\item{\code{visibility}}{The visibility CSS attribute.}

\item{\code{width}}{The width CSS attribute.}

\item{\code{...}}{additional parameters for \link{jupyter.widget.Widget}}

\item{\code{error_call}}{See \code{\link[rlang:args_error_context]{rlang::args_error_context()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-jupyter.widget.Layout-clone"></a>}}
\if{latex}{\out{\hypertarget{method-jupyter.widget.Layout-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{jupyter.widget.Layout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
