% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Layout.R
\name{Layout}
\alias{Layout}
\title{Layout}
\usage{
Layout(..., error_call = current_env())
}
\arguments{
\item{...}{See the constructor for \link{jupyter.widget.Layout}}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a new \code{jupyter.widget.Layout} object
}
\description{
Layout
}
